+++ lib/src/annotations/sql_index.dart
@@ -1,0 +1,11 @@
+/// define index on table
+class SqlIndex {
+  /// name for index
+  final String name;
+
+  /// columns for index
+  final List<String> columns;
+
+  /// constructor
+  const SqlIndex({required this.name, required this.columns});
+}